fastapi_solution/requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
elastic-transport==8.4.0
elasticsearch==7.9.1
fastapi==0.88.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hiredis==2.0.0
httptools==0.5.0
idna==3.4
multidict==6.0.2
orjson==3.8.2
packaging==21.3
psycopg2-binary==2.9.5
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
redis==4.3.5
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.1
